fix wasm2c tests on 32-bit (#2124)
The spectest_make_externref() in test/spec-wasm2c-prefix.c:203 assumed
that u64 (uint64_t) can be converted to wasm_rt_externref_t (void *).

+++ test/spec-wasm2c-prefix.c
@@ -200,7 +200,7 @@ static bool is_equal_wasm_rt_funcref_t(wasm_rt_funcref_t x,
          (x.module_instance == y.module_instance);
 }
 
-wasm_rt_externref_t spectest_make_externref(u64 x) {
+wasm_rt_externref_t spectest_make_externref(uintptr_t x) {
   return (wasm_rt_externref_t)(x + 1);  // externref(0) is not null
 }
